applications/flask_rest/requirements.txt
Flask == 3.0.1
gunicorn == 21.2.0
